Add support for two GPUs
Some notebooks habe hybrid graphics and therefor multiple GPUs.
This adds a function to find GPUs in the pci device list. An index is passed to determine which GPU should be returned. If no GPU can be found for a given index, an empty string is returned.
Similiar to color1 and color2, two functions called get_gpu1 and get_gpu2 are added. The configuration now has two GPU entries, but if the seconds entry contains an empty string, it is not printed to the terminal.
Another pci class name had to be added for identification of GPUs, as my Toshiba notebook reports his NVIDIA GPU as a "3D controller" and not as a "VGA compatible controller".